# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 隐私邮箱外部取件 API v2
  version: 2.0.0
  description: |
    本文档主要描述外部接口，并记录升级所需的少量管理兼容接口。服务提供两种
    取件能力：标准只读 IMAPS，以及使用 API Key 或派生取码 URL 读取验证码历史。

    管理端可为每个主号配置上游隐式 TLS IMAP 主机和端口；创建时默认使用
    `imap.mail.me.com:993`。修改规范化后的主机、端口或用户名代表切换邮箱来源，会清理
    旧来源的同步游标、兼容快照、消费历史、待处理 `\\Seen` 任务、v2 归档和 OTP 历史，
    轮换公开 IMAPS 的 `UIDVALIDITY`，并将同步状态重置为 `pending`。

    主号也可以使用 `custom` 邮箱模式并保存邮箱后缀。自定义随机邮箱使用 8–12 位
    ASCII 小写英文字母和数字作为本地部分，批量内和数据库中均禁止重复；该流程不调用 Apple，原 iCloud
    自动创建计划和限流规则保持不变。

    收件路由有三套：`mailbox_type=icloud` 加默认 `imap.mail.me.com:993` 是直连 iCloud IMAP；
    `mailbox_type=icloud` 加非默认 IMAP 来源（主机/端口或用户名不同于 iCloud 默认）是 iCloud 隐私邮箱转发到第三方 IMAP；
    `mailbox_type=custom` 是自定义域名邮箱，按原始投递收件人头匹配本地地址。

    如果 iCloud 隐私邮箱转发到第三方邮箱，仍应保持 `mailbox_type=icloud`，并将上游
    `imap_username` 填为第三方 IMAP 登录/最终物理投递目标地址（若该邮箱服务允许使用邮箱地址登录）。同步会优先使用
    `X-ICLOUD-HME` 的 `p=`/`f=` 路由信息，再结合 `Delivered-To`、`X-Original-To`、
    `Original-Recipient` 和常见 Apple 原始收件人头恢复隐私邮箱归属；Apple 的 `f=` 可以是中间
    转发地址，不必与最终 `Delivered-To` 相同。若只在 `To/Cc` 留下隐私地址，必须显式启用弱收件人头才会匹配。
    这些规则只控制邮件取件与归属；Apple 目录同步和自动创建仍使用现有的主号默认转发目标校验。

    IMAPS 用户名是隐私邮箱地址。客户端可以使用本服务签发的 IMAP 密码登录，
    或通过 `POST /oauth2/v2.0/token` 获取一小时访问令牌后使用 XOAUTH2。
    邮箱只有只读 INBOX，支持 LIST、STATUS、SELECT/EXAMINE、SEARCH、
    FETCH/UID FETCH 和 IDLE；所有写操作返回只读错误。

    新 v2 alias 的 API Key、IMAP 密码、client ID、refresh token 与 access token
    均由本服务签发。legacy alias 的兼容轮换只替换 API Key 和由其派生的旧直达链接，
...
    管理员可显式执行全量凭据轮换：所有 legacy alias 会强制迁移为 v2，所有现有
...
servers:
...
paths:
  /api/v1/pool/leases/{leaseID}/code:
    get:
      tags: [Mailbox Pool]
      operationId: readPoolLeaseCode
      summary: 读取项目所领取邮箱的最新验证码，可选有界等待
      description: |
        Pool Key 必须属于 lease 的项目。lease 为有效 leased 或 used，且主号和 alias 启用。
        查询起点取 after 与领取时间较晚者，只返回严格晚于起点的验证码。
        wait_seconds 默认 0，保留原单次按需读取；启用共享收件时正值允许等待新归档，
        预算包含首轮同步。正常到期仍返回 no_code，首轮同步未完成或上游错误返回 503。
        未配置共享通知时退回单次读取。等待不持有全局轮换/Pool 锁，最终重新校验归属及状态。
        共享模式同 root 容量 2，每 2 秒恢复 1 次额度；最多一个普通读取和一个长等待，
        等待期间额度恢复，完成后不追加冷却。旧非共享路径保持独占和完成后 2 秒冷却。
        超额返回 429；被拦截请求不延长恢复时间。
      security:
        - poolKey: []
      parameters:
        - name: leaseID
          in: path
          required: true
          schema: {type: string}
        - name: after
          in: query
          schema: {type: string, format: date-time}
        - name: wait_seconds
          in: query
          description: 单个十进制整数；默认 0。HTTP 超时应留网络收尾余量。
          schema: {type: integer, minimum: 0, maximum: 15, default: 0}
      responses:
        "200":
          description: 命中新验证码或本次等待范围内无新码；不返回邮箱凭据。
          headers:
            Cache-Control:
              schema: {type: string, const: no-store}
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    oneOf:
                      - type: object
                        required: [success, otp, time, email]
                        properties:
                          success: {type: boolean, const: true}
                          otp: {type: string}
                          time: {type: string, format: date-time}
                          email: {type: string, format: email}
                      - type: object
                        required: [success, code, retryable, email]
                        properties:
                          success: {type: boolean, const: false}
                          code: {type: string, const: no_code}
                          retryable: {type: boolean, const: true}
                          email: {type: string, format: email}
        "400":
          description: INVALID_POOL_REQUEST，参数非法或 wait_seconds 重复。
        "401":
          description: INVALID_POOL_KEY，项目 Key 无效、已轮换或停用。
        "404":
          description: NOT_FOUND，lease 不存在或属于其他项目。
        "409":
          description: LEASE_CLOSED，领取关闭、过期或主号/alias 停用。
        "429":
          description: RATE_LIMITED，同类取件在途、速率预算不足或容量已满。在途/容量原因返回建议复查间隔。
          headers:
            Retry-After:
              description: 剩余等待时间向上取整的秒数。
              schema: {type: string, pattern: '^[1-9][0-9]*$'}
            X-Retry-After-Ms:
              description: 剩余等待时间或建议复查间隔，单位毫秒。
              schema: {type: string, pattern: '^[1-9][0-9]*$'}
            X-RateLimit-Reason:
              schema:
                type: string
                enum: [mailbox_rate, mailbox_waiting, mailbox_inflight, mailbox_cooldown, account_concurrency, global_concurrency, global_rate, mailbox_capacity]
        "503":
          description: SYNC_UNAVAILABLE 或 POOL_UNAVAILABLE；同步异常时 Retry-After 为 3。
...
  /api/v1/mail/latest:
    get:
      tags:
        - Legacy Compatibility
      operationId: getLegacyLatestMail
      summary: 兼容读取当前最新完整邮件快照
      description: |
        使用原有 `Authorization: Bearer <API Key>`。请求可重复执行，不消费邮件，
        不写入消费记录，也不改变 IMAP `Seen` 状态；已领取 legacy Key 和 v2 Key
        均可按绑定 alias 读取。
      security:
        - aliasAPIKey: []
      responses:
        "200":
          description: 当前最新邮件快照
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyLatestMailResponse"
        "401":
          $ref: "#/components/responses/InvalidCredential"
        "404":
          $ref: "#/components/responses/MailNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/mail/recent:
    get:
      tags:
        - Legacy Compatibility
      operationId: getLegacyRecentMail
      summary: 兼容消费最近一小时内的当前邮件
      description: |
        使用原有 `api_key` 查询参数（只能出现一次），接受 API Key、legacy alias 的
        原 v1 直达 token，或 v2 alias 的 recent-mail 专用 token。成功后原子记录消费
        状态并异步写入主号 IMAP `Seen`；同一 UID 的后续请求继续返回
        `MAIL_NOT_FOUND`。旧 legacy 直达链接保持路径和 token 不变；OTP 专用 token
        不授权此接口。
      security:
        - aliasAPIKeyQuery: []
      parameters:
        - name: api_key
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 当前快照已消费并加入 IMAP Seen 队列
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacyRecentMailResponse"
        "401":
          $ref: "#/components/responses/InvalidCredential"
        "404":
          $ref: "#/components/responses/MailNotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/otp:
    get:
      summary: 读取当前隐私邮箱的验证码历史
      description: |
        默认返回最新优先、持续累积的裸数组，每个邮箱最多 100 条。设置
        `ICLOUD_API_OTP_RETURN_LATEST_ONLY=true` 后，有记录时只返回该数组第一项的
        裸对象，也就是真正最新的有效 OTP。旧变量 `ICLOUD_API_OTP_RETURN_LAST_ONLY`
        保留为兼容别名；同时设置时以正式变量为准。
        重复读取不会消费验证码，也不会修改邮件已读状态。两种模式没有验证码时均返回 `200 []`。
        调用方必须二选一：使用 Bearer API Key，或传入管理端签发取码 URL 中的 OTP
        专用派生 token。迁移或轮换后 URL 路径仍为 `/api/v1/otp`；v2 alias 的 OTP
        token 保持用途隔离的 v3 公开信封版本，与 recent-mail 的 v2 信封分别签名。
        这里的 v3 只表示 token 信封版本，不表示新的 alias credential mode。
        recent-mail token 和原 v1 直达 token 不授权此接口。
      security:
        - aliasAPIKey: []
        - derivedToken: []
      parameters:
        - name: token
          in: query
          required: false
          description: 取码 URL 中的 OTP 专用派生 token；使用时不得同时发送 Authorization。
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: 默认是验证码历史数组；最新单条模式有记录时为对象；没有记录时为一个空数组。
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/OTPRecord"
                  - $ref: "#/components/schemas/OTPRecord"
              examples:
                records:
                  value:
                    - otp: "876543"
                      time: "2026-08-11T12:00:00+08:00"
                    - otp: "123456"
                      time: "2026-08-11T11:00:00+08:00"
                latestOnly:
                  value:
                    otp: "876543"
                    time: "2026-08-11T12:00:00+08:00"
                empty:
                  value: []
        "401":
          $ref: "#/components/responses/InvalidCredential"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  securitySchemes:
    externalOAuthToken:
      type: http
      scheme: bearer
      bearerFormat: deployment OAuth token
    aliasAPIKey:
      type: http
      scheme: bearer
      bearerFormat: service-issued API Key
    derivedToken:
      type: apiKey
      in: query
      name: token
      description: 仅授权只读 OTP 历史；不能用于消费 recent-mail。
    aliasAPIKeyQuery:
      type: apiKey
      in: query
      name: api_key
    poolKey:
      type: http
      scheme: bearer
      bearerFormat: pool project Key
      description: 以 `pool_` 开头的项目 Key；仅与管理员 Session 组合时可调用 Pool 已使用邮箱退休删除接口。
    adminSession:
      type: apiKey
      in: cookie
      name: icloud_admin_session
...
  schemas:
...
    OTPRecord:
      type: object
      additionalProperties: false
      required: [otp, time]
      properties:
        otp:
          type: string
          pattern: "^[0-9]{6}$"
        time:
          type: string
          format: date-time
...
    LegacyLatestMailResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: true
          required: [alias, message, synced_at, stale]
          properties:
            alias:
              type: string
              format: email
            message:
              $ref: "#/components/schemas/LegacyMailMessage"
            synced_at:
              type: string
              format: date-time
            stale:
              type: boolean
    LegacyRecentMailResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [address, subject, snippet, sent_at]
          properties:
            address:
              type: string
              format: email
            subject:
              type: string
            snippet:
              type: string
            sent_at:
              type: string
              format: date-time
    LegacyMailMessage:
      type: object
      additionalProperties: true
      required: [id, message_id, received_at, from, to, cc, subject, text, html, attachments, has_attachments, body_truncated]
      properties:
        id:
          type: string
        message_id:
          type: string
        received_at:
          type: string
          format: date-time
        sent_at:
          type:
            - string
            - "null"
          format: date-time
        from:
          type: array
          items:
            $ref: "#/components/schemas/MailAddress"
        to:
          type: array
          items:
            $ref: "#/components/schemas/MailAddress"
        cc:
          type: array
          items:
            $ref: "#/components/schemas/MailAddress"
        subject:
          type: string
        text:
          type: string
        html:
          type: string
        attachments:
          type: array
          items:
            type: object
            additionalProperties: true
        has_attachments:
          type: boolean
        body_truncated:
          type: boolean
    MailAddress:
      type: object
      additionalProperties: true
      required: [email]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
...
    APIErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
  responses:
...
    InvalidCredential:
      description: Bearer 凭据、API Key 或派生 token 无效。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
...
    MailNotFound:
      description: 当前没有符合条件且尚未消费的邮件（`MAIL_NOT_FOUND`）。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
...
    RateLimited:
      description: 请求过于频繁。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
    Unavailable:
      description: 服务暂不可用。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIErrorEnvelope"
...